requirements.txt
Faker==21.0.0
pandas==2.1.4
pandas_gbq==0.20.0
protobuf==4.25.1
google-cloud-bigquery==3.14.1
